backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Supabase
supabase==2.7.4
postgrest==0.16.8

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Clients
httpx==0.25.2
aiofiles==23.2.1

# LLM Clients (from original app)
openai==1.40.8
anthropic==0.37.1
groq==0.9.0

# NLP and Processing (from original app)
spacy==3.7.2
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS
python-cors==1.0.1

# Logging
structlog==23.2.0
